+++ global_trotter_web/src/App.jsx
@@ -3,6 +3,7 @@ import ItineraryDraftProvider from './context/ItineraryDraftProvider.jsx'
 import Landing from './pages/Landing.jsx'
 import Register from './pages/Register.jsx'
 import Login from './pages/Login.jsx'
+import SelectStyle from './pages/SelectStyle.jsx'
 import Home from './pages/Home.jsx'
 import Destinations from './pages/Destinations.jsx'
 import DestinationDetails from './pages/Destinationdetails.jsx'
@@ -19,6 +20,7 @@ function App() {
           <Route path="/" element={<Landing />} />
           <Route path="/register" element={<Register />} />
           <Route path="/login" element={<Login />} />
+          <Route path="/select-style" element={<SelectStyle />} />
           <Route path="/home" element={<Home />} />
           <Route path="/destinations" element={<Destinations />} />
           <Route path="/destinations/:id" element={<DestinationDetails />} />